requirements.txt
gradio==3.50.2
python-decouple==3.8
langchain==0.0.312
memoization==0.4.0
chromadb==0.4.3
cohere==4.27
openai==0.27.8
lark==1.1.7
domdf_python_tools==3.7.0